Component pattern works
