+++ contracts/contract-config.js
@@ -5,8 +5,8 @@ module.exports = {
     verifierAddress: '0xbbbb93A6B3A1D6fDd27909729b95CCB0cc9002C0',
     blacklisterAddress: null,
     bootstrapSPIds: [1,2,3],
-    bootstrapSPDelegateWallets: ['0xc2542A086011Ae797667964460e33B10f0355A0C', '0x7F68e2E5CfD0200a096C87d1F1A3991d6fb421dF', '0x19AB495435Da2A463dA0B2eF5aD12E9438ab8279'],
-    bootstrapSPOwnerWallets: ['0xc2542A086011Ae797667964460e33B10f0355A0C', '0x7F68e2E5CfD0200a096C87d1F1A3991d6fb421dF', '0x19AB495435Da2A463dA0B2eF5aD12E9438ab8279'],
+    bootstrapSPDelegateWallets: ['0x203fc83304a11003b4c845B3078c2D24df02fd9b', '0x4839c9d2046d7862fcc2D029755e28c4528C2946', '0xe1Af216be3Ca7c520BB69310547a1AB9ad30872e'],
+    bootstrapSPOwnerWallets: ['0x203fc83304a11003b4c845B3078c2D24df02fd9b', '0x4839c9d2046d7862fcc2D029755e28c4528C2946', '0xe1Af216be3Ca7c520BB69310547a1AB9ad30872e'],
     userReplicaSetBootstrapAddress: null
   },
   test_local: {
