+++ creator-node/src/app.js
@@ -18,7 +18,7 @@ const {
 const config = require('./config')
 const healthCheckRoutes = require('./components/healthCheck/healthCheckController')
 const contentBlacklistRoutes = require('./components/contentBlacklist/contentBlacklistController')
-const utils = require('./utils')
+const replicaSetRoutes = require('./components/replicaSet/replicaSetController')
 
 const app = express()
 // middleware functions will be run in order they are added to the app below
@@ -41,6 +41,7 @@ app.use(getRateLimiterMiddleware())
 require('./routes')(app)
 app.use('/', healthCheckRoutes)
 app.use('/', contentBlacklistRoutes)
+app.use('/', replicaSetRoutes)
 
 function errorHandler (err, req, res, next) {
   req.logger.error('Internal server error')
