+++ discovery-provider/compose/.env
@@ -12,4 +12,4 @@ audius_delegate_private_key=0x3873ed01bfb13621f9301487cc61326580614a5b99f3c33cf3
 audius_discprov_identity_service_url=http://audius-identity-service_identity-service_1:7000
 audius_discprov_user_metadata_service_url=http://cn1_creator-node_1:4000
 audius_discprov_blacklist_block_indexing_interval=5
-audius_eth_contracts_registry=0x3BB75Bb7aef46d3C2A05EA4B64BbFeC08F7feE0F
+audius_eth_contracts_registry=0x76A814d95b3bcdd4B28DC07663D2Db4Bbbc8c5EB
